change component + delete console.logs
