+++ src/app/main/component/auth/components/sign-in/GoogleOAuthProvider/GoogleOAuthProvider.ts
@@ -2,7 +2,5 @@ import { GoogleOAuthProvider } from 'google-oauth-gsi';
 import { environment } from '@environment/environment';
 
 export const googleProvider = new GoogleOAuthProvider({
-  clientId: environment.googleClientId,
-  onScriptLoadError: () => console.error('Error loading GoogleOAuth scripts'),
-  onScriptLoadSuccess: () => console.log('GoogleOAuth scripts are loaded successfully')
+  clientId: environment.googleClientId
 });
