+++ src/app/ubs/ubs-admin/components/ubs-admin-table/table-cell-readonly/table-cell-readonly.component.scss
@@ -47,11 +47,3 @@ div {
 .pointer {
   cursor: pointer;
 }
-
-.pencil {
-  @include pencil();
-}
-
-.padding-right {
-  padding-right: 25px;
-}
