+++ src/app/ubs/ubs-admin/components/ubs-admin-table/table-cell-readonly/table-cell-readonly.component.ts
@@ -1,15 +1,11 @@
-import { Component, DestroyRef, inject, Input, OnChanges, OnInit } from '@angular/core';
-import { Address, IBigOrderTableOrderInfo, IColumnBelonging } from '../../../models/ubs-admin.interface';
+import { Component, inject, Input, OnChanges, OnInit } from '@angular/core';
+import { IColumnBelonging } from '../../../models/ubs-admin.interface';
 import { Language } from 'src/app/main/i18n/Language';
 import { TableKeys } from '../../../services/table-keys.enum';
 import { Patterns } from 'src/assets/patterns/patterns';
 import { PaymnetStatus } from 'src/app/ubs/ubs/order-status.enum';
 import { AdminTableService } from '@ubs/ubs-admin/services/admin-table.service';
-import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
-import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
-import { UBSAddAddressPopUpComponent } from 'src/app/shared/ubs-add-address-pop-up/ubs-add-address-pop-up.component';
-import { OrderService } from '@ubs/ubs-admin/services/order.service';
-import { pipe, switchMap } from 'rxjs';
+
 @Component({
   selector: 'app-table-cell-readonly',
   templateUrl: './table-cell-readonly.component.html',
@@ -27,10 +23,8 @@ export class TableCellReadonlyComponent implements OnInit, OnChanges {
   dataObj: IColumnBelonging = null;
   data: string | number | { ua: string; en: string } | null;
   private readonly font = '12px Lato, sans-serif';
-  destroyRef = inject(DestroyRef);
-  dialog = inject(MatDialog);
   adminTableService = inject(AdminTableService);
-  orderService = inject(OrderService);
+
 
   ngOnInit(): void {
     if (this.optional?.length) {
@@ -57,9 +51,7 @@ export class TableCellReadonlyComponent implements OnInit, OnChanges {
         const { regex, match } = replaceRules[this.lang];
         this.title = (this.title as string).toLowerCase().replace(regex, (el) => match[el]);
       }
-
       this.data = this.title;
-
       this.isStatus();
     }
   }
@@ -82,29 +74,5 @@ export class TableCellReadonlyComponent implements OnInit, OnChanges {
 
   onMouseEnter(event: MouseEvent, tooltip: any): void {
     this.adminTableService.showTooltip(event, tooltip, this.font);
-  }
-
-  isAddressKey(): boolean {
-    return this.key === 'region' || this.key === 'city' || this.key === 'district' || this.key === 'address';
-  }
-
-  openEditAddressWindow(): void {
-    this.orderService
-      .getOrderInfo(this.orderId)
-      .pipe(
-        switchMap((orderInfo) => {
-          const dialogConfig = new MatDialogConfig();
-          dialogConfig.panelClass = 'address-matDialog-styles';
-          dialogConfig.data = {
-            edit: true,
-            addFromProfile: true,
-            address: orderInfo.addressExportDetailsDto
-          };
-          const dialogRef = this.dialog.open(UBSAddAddressPopUpComponent, dialogConfig);
-          return dialogRef.afterClosed();
-        }),
-        takeUntilDestroyed(this.destroyRef)
-      )
-      .subscribe(() => {});
   }
 }
