+++ src/app/ubs/ubs-admin/components/ubs-admin-table/ubs-admin-table.component.ts
@@ -17,7 +17,6 @@ import {
   AddFilterMultiAction,
   AddFiltersAction,
   ChangingOrderData,
-  ClearFilters,
   GetColumns,
   GetColumnToDisplay,
   GetLocationsDetails,
